1) Create separate local directories for each image with datasore authentication information
2) Added support to store Image UUID for all images.
   - Use this for any image name uses instead of safe name.

3) Fixed a bug to not remove rkt files un-intentionally

+++ pkg/pillar/cmd/domainmgr/domainmgr.go
@@ -1276,6 +1276,9 @@ func doInactivate(ctx *domainContext, status *types.DomainStatus) {
 	// Do we need to delete any rw files that should
 	// not be preserved across reboots?
 	for _, ds := range status.DiskStatusList {
+		if status.IsContainer {
+			continue
+		}
 		if !ds.ReadOnly && !ds.Preserve {
 			log.Infof("Delete copy at %s\n", ds.ActiveFileLocation)
 			if err := os.Remove(ds.ActiveFileLocation); err != nil {
